+++ resources/js/components/StanzaView.vue
@@ -1,50 +1,38 @@
 <template>
   <div>
-    <div class="row mb-3">
-      <div class="col-md-2">Name</div>
-      <div class="col-md-10">
-        {{ entry.name }}
-      </div>
-    </div>
-    <div v-if="entry.display_name" class="row mb-3">
-      <div class="col-md-2">Display Name</div>
-      <div class="col-md-10">
-        {{ entry.display_name }}
-      </div>
-    </div>
-    <div v-if="entry.comment" class="row mb-3">
-      <div class="col-md-2">Comment</div>
-      <div class="col-md-10">
-        <pre style="white-space: pre-wrap">{{ Array.isArray(entry.comment) ? entry.comment.join("\n") : entry.comment }}</pre>
-      </div>
-    </div>
-    <div v-if="stanza.hasOwnProperty('oclcIncludeFile')" class="row mb-3">
-      <div class="col-md-2">OCLC Hosted Include File</div>
-      <div class="col-md-10">
-        {{ stanza.oclcIncludeFile }}
-      </div>
-    </div>
-    <div class="row mb-3">
-      <div class="col-md-2">Status</div>
-      <div class="col-md-10">
-        <template v-if="entry.active == false">
-          <strong>Inactive</strong>. This stanza will still appear in your EZproxy config, but will be marked inactive (i.e. treated like a comment).
-        </template>
-        <template v-else>
-          Active
-        </template>
-      </div>
-    </div>
+    <b-card no-body>
+      <b-tabs pills card vertical end>
+        <b-tab title="Details" active>
+          <b-card-text>
+            <StanzaViewDetails :entry="entry" :stanza="stanza"></StanzaViewDetails>
+          </b-card-text>
+        </b-tab>
+        <b-tab title="Rules">
+          <b-card-text>
+            <StanzaViewRules :filename="filename" :entry="entry" :stanza="stanza"></StanzaViewRules>
+          </b-card-text>
+        </b-tab>
+      </b-tabs>
+    </b-card>
   </div>
 </template>
 
 <script>
+import StanzaViewDetails from './StanzaViewDetails';
+import StanzaViewRules from './StanzaViewRules';
+
 export default {
   name: 'StanzaView.vue',
 
+  components: {
+    StanzaViewDetails,
+    StanzaViewRules
+  },
+
   props: {
-    entry: Object,
-    stanza: Object
+    filename: { type: String, required: true },
+    entry: { type: Object, required: true },
+    stanza: { type: Object, default: () => ({}) }
   }
 }
 </script>